some additional changes to support icpc on ia64
git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/trunk@286 4eea3252-f0fb-4393-894d-40516dce545b

+++ src/sgpp/grid/storage/hashmap/HashGridIndex.hpp
@@ -431,7 +431,7 @@ public:
     		}
     		else
     		{
-    			return_stream << (std::pow(0.5, level[i])*index[i]);
+    			return_stream << (pow(0.5, static_cast<double>(level[i]))*index[i]);
     		}
 
     		if (i < DIM-1)
